Make the sync settings available to the graphql/service layer
This is needed to get the central server url to do the central user
login call.
