+++ service/src/lib.rs
@@ -25,6 +25,7 @@ pub mod stocktake;
 pub mod stocktake_line;
 pub mod store;
 pub mod sync_processor;
+pub mod sync_settings;
 pub mod token;
 pub mod token_bucket;
 pub mod user_account;
